implement one global param (favorite language)
