+++ src/utils/consts/settings.js
@@ -2,6 +2,7 @@ const settings = {
   keys: {
     global: 'global',
     deck: 'deck',
+    favoriteLanguage: { global: true, key: 'favoriteLanguage' },
     typeGrouping: { global: false, key: 'typeGrouping' },
     typePriority: { global: false, key: 'typePriority' },
     sorting: { global: false, key: 'sorting' },
